% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsopts_default.R
\name{dsopt_default}
\alias{dsopt_default}
\title{Retrieve Default Option}
\usage{
dsopt_default(opt)
}
\arguments{
\item{opt}{A character string specifying the name of the option whose default value is to be retrieved.}
}
\value{
The default value of the specified option. If the option does not exist, returns NULL.
}
\description{
This function retrieves the default value for a specified option from the \code{dsopts} package.
}
\examples{
\dontrun{
# Assuming 'my_option' as a valid option name
default_opt <- dsopt_default(opt = 'my_option')
}
}
