% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsopts-package.R
\docType{package}
\name{dsopts-package}
\alias{dsopts}
\alias{dsopts-package}
\title{dsopts: dsopts}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Camila Achuri \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
